Display implementation
Created the Display class for handling the basic displaying behaviour for the game. Now the game is rendered nicely in fixed NxN resolution independent from the current window size and / or if it's maximized. In the meantime, I've added the raygui library to the project but I'm thinking now about getting rid of it and handling text rendering for my own.