Background addition
Added the scrollable background to the game. What's more, I've removed the raygui library, because I was only using it for text labels, and I replaced them with the casual text rendering privided by raylib-cpp. In the gameplay scene there is now a little tooltip informing the player how to start the game. For another commit I'll probably replace all the filtering for something less blurry, add the scoring and HP and maybe workaround some other things.