HP and Score
From now on, the HP and the Score will be rendered to the screen! Also, the audio device was prepared in the main() so in the near future I'll work on SFX. Also, I need to find a way to nerf the jumping on the floor.. In the meantime, the delay between dropping the eggs was modified so on every frame the delay is decreased by the delta time multiplied by some factor.