Final Easter Hunt 1.0.0 commit
Getting closer to release of the Easter Hunt. All the content that was meant to be there was added. Now the only things to do is to create a README, play around with build system to create the fnal executables for both Linux and Windows and call it a day! <3